+++ controller/doctorController.js
@@ -2,7 +2,7 @@ const Doctor = require("../models/doctor");
 const express = require("express");
 const exchangeRequestModel = require("../models/exchangeRequest")
 const mongoose = require("mongoose");
-
+const Ward=require('../models/ward')
 
 const getUser = async (req, res) => {
   const doctorList = await Doctor.find();
@@ -149,6 +149,106 @@ const getShiftNames=(req,res)=>{
   return res.status(200).json({"shiftNames":shiftNames});
 }
 
+const getUserDetails = async (req, res) => {
+  const userName = req.body.userName;
+  const userType = req.body.type;
+  console.log(req.body);
+  const a = null;
+  let userDetails = null;
+  let wardDetails = null;
+  console.log(userName);
+
+  if (userType === "1") {
+    userDetails = await Doctor.findOne({ emailaddress: userName });
+    if (!userDetails) {
+      console.log("doctor not found");
+      return res.status(500).json({
+        success: false,
+        msg: "errroorrrrrrrrrr",
+      });
+    } else {
+      console.log(userDetails);
+      wardDetails = await Ward.findOne({ id: userDetails.wardID });
+      console.log(wardDetails);
+      console.log(wardDetails.wardNumber);
+      return res.status(200).json({
+        success: true,
+        msg: "get doctor profile details correctly",
+        fullName: userDetails.firstName + " " + userDetails.lastName,
+        email: userDetails.emailaddress,
+        address: userDetails.address,
+        telephone: userDetails.telephone,
+        emailaddress: userDetails.emailaddress,
+        userName: userDetails.userName,
+        wardName: wardDetails.wardName,
+        wardID: wardDetails.wardNumber,
+      });
+      // res.send(userDetails);
+    }
+  }
+
+  // get consultant details
+  else if (userType === "2") {
+    userDetails = await Consultant.findOne({ emailaddress: userName });
+    if (!userDetails) {
+      console.log("consultant not found");
+      return res.status(500).json({
+        success: false,
+        msg: "errroorrrrrrrrrr",
+      });
+    } else {
+      console.log(userDetails);
+      wardDetails = await Ward.findOne({ id: userDetails.wardID });
+      console.log(wardDetails.wardName);
+      return res.status(200).json({
+        success: true,
+        msg: "get consultant profile details correctly",
+        fullName: userDetails.firstName + " " + userDetails.lastName,
+        email: userDetails.emailaddress,
+        address: userDetails.address,
+        telephone: userDetails.telephone,
+        emailaddress: userDetails.emailaddress,
+        wardName: wardDetails.wardName,
+        wardId: wardDetails.wardNumber,
+        userName: userDetails.userName,
+        speciality: userDetails.speciality,
+      });
+      // res.send(userDetails);
+    }
+  }
+
+  // get admin details
+  else if (userType === "3") {
+    userDetails = await Admin.findOne({ emailaddress: userName });
+    if (!userDetails) {
+      console.log("consultant not found");
+      return res.status(500).json({
+        success: false,
+        msg: "errroorrrrrrrrrr",
+      });
+    } else {
+      console.log(userDetails);
+      wardDetails = await Ward.findOne({ id: userDetails.wardID });
+      console.log(wardDetails);
+      return res.status(200).json({
+        success: true,
+        msg: "get admin profile details correctly",
+        fullName: userDetails.firstName + " " + userDetails.lastName,
+        email: userDetails.emailaddress,
+        address: userDetails.address,
+        telephone: userDetails.telephone,
+        emailaddress: userDetails.emailaddress,
+        wardName: wardDetails.wardName,
+        wardId: wardDetails.wardNumber,
+        userName: userDetails.userName,
+        speciality: userDetails.speciality,
+      });
+      // res.send(userDetails);
+    }
+  }
+};
+
+
 module.exports = {
-  getUser,getData,submitLeaveRequest,submitPreferrableSlots,getIndividualRoster,getShiftNames, getInNotif, putNotif, getOutNotif, hideNotif, declineRequest, acceptRequest
+  getUser,getData,submitLeaveRequest,submitPreferrableSlots,getIndividualRoster,getShiftNames, getInNotif, putNotif, getOutNotif, hideNotif, declineRequest, acceptRequest,getUserDetails
 };
