+++ routes/userAdminRouter.js
@@ -1,9 +1,9 @@
 const express=require("express");
 const router=express.Router();
 const adminController=require("../controller/adminController")
-const userController=require("../controller/userController")
+
 router.get('/getUser',adminController.getUser)
 router.all('/addUser',adminController.addUser)
 
-router.post('/userDetails',userController.getUserDetails)
+router.post('/userDetails',adminController.getUserDetails)
 module.exports=router;
