+++ routes/userConsultantRouter.js
@@ -1,8 +1,8 @@
 const express=require("express");
 const router=express.Router();
 const consultantController=require("../controller/consultantController")
-const userController=require("../controller/userController")
+
 router.get('/getUser',consultantController.getUser)
 
-router.post('/userDetails',userController.getUserDetails)
+router.post('/userDetails',consultantController.getUserDetails)
 module.exports=router;
